requirements.txt
pipelinewise-singer-python==2.0.1
chardet==4.0.0
paramiko==2.6.0
backoff==1.11.1
terminaltables==3.1.0
python-gnupg @ git+ssh://git@github.com/symon-ai/python-gnupg.git@0.4.9.1#egg=python-gnupg
file-processors @ git+ssh://git@github.com/symon-ai/file-processors.git@main
boto3==1.24.35
smart_open==6.1.0